Extract ConfigurationCacheTaskSerializationIntegrationTest
from ConfigurationCacheIntegrationTest
