+++ subprojects/configuration-cache/src/integTest/groovy/org/gradle/configurationcache/ConfigurationCacheTaskSerializationIntegrationTest.groovy
@@ -1,0 +1,241 @@
+/*
+ * Copyright 2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.configurationcache
+
+import spock.lang.Unroll
+
+import javax.inject.Inject
+
+class ConfigurationCacheTaskSerializationIntegrationTest extends AbstractConfigurationCacheIntegrationTest {
+
+    def "restores task fields whose value is an object graph with cycles"() {
+        buildFile << """
+            class SomeBean {
+                String value
+                SomeBean parent
+                SomeBean child
+
+                SomeBean(String value) {
+                    println("creating bean")
+                    this.value = value
+                }
+            }
+
+            class SomeTask extends DefaultTask {
+                @Internal
+                final SomeBean bean
+
+                SomeTask() {
+                    bean = new SomeBean("default")
+                    bean.parent = new SomeBean("parent")
+                    bean.parent.child = bean
+                    bean.parent.parent = bean.parent
+                }
+
+                @TaskAction
+                void run() {
+                    println "bean.value = " + bean.value
+                    println "bean.parent.value = " + bean.parent.value
+                    println "same reference = " + (bean.parent.child == bean)
+                }
+            }
+
+            task ok(type: SomeTask) {
+                bean.value = "child"
+            }
+        """
+
+        when:
+        configurationCacheRun "ok"
+
+        then:
+        result.output.count("creating bean") == 2
+
+        when:
+        configurationCacheRun "ok"
+
+        then:
+        outputDoesNotContain("creating bean")
+        outputContains("bean.value = child")
+        outputContains("bean.parent.value = parent")
+        outputContains("same reference = true")
+    }
+
+    def "replaces provider with fixed value"() {
+        buildFile << """
+            class SomeTask extends DefaultTask {
+                @Internal
+                Provider<String> value
+
+                @TaskAction
+                void run() {
+                    println "this.value = " + value.getOrNull()
+                }
+            }
+
+            task ok(type: SomeTask) {
+                value = providers.provider {
+                    println("calculating value")
+                    'value'
+                }
+            }
+        """
+
+        when:
+        configurationCacheRun "ok"
+
+        then:
+        outputContains("calculating value")
+        outputContains("this.value = value")
+
+        when:
+        configurationCacheRun "ok"
+
+        then:
+        outputDoesNotContain("calculating value")
+        outputContains("this.value = value")
+    }
+
+    @Unroll
+    def "Directory value can resolve paths after being restored"() {
+        buildFile << """
+            import ${Inject.name}
+
+            class SomeTask extends DefaultTask {
+                @Internal
+                Directory value
+                @Internal
+                final Property<Directory> propValue
+
+                @Inject
+                SomeTask(ObjectFactory objects) {
+                    propValue = objects.directoryProperty()
+                }
+
+                @TaskAction
+                void run() {
+                    println "value = " + value
+                    println "value.child = " + value.dir("child")
+                    println "propValue = " + propValue.get()
+                    println "propValue.child = " + propValue.get().dir("child")
+                    println "propValue.child.mapped = " + propValue.dir("child").get()
+                }
+            }
+
+            task ok(type: SomeTask) {
+                value = layout.projectDir.dir("dir1")
+                propValue = layout.projectDir.dir("dir2")
+            }
+        """
+
+        when:
+        configurationCacheRun "ok"
+        configurationCacheRun "ok"
+
+        then:
+        outputContains("value = ${file("dir1")}")
+        outputContains("value.child = ${file("dir1/child")}")
+        outputContains("propValue = ${file("dir2")}")
+        outputContains("propValue.child = ${file("dir2/child")}")
+        outputContains("propValue.child.mapped = ${file("dir2/child")}")
+    }
+
+    def "restores task abstract properties"() {
+        buildFile << """
+            interface Bean {
+                @Internal
+                Property<String> getValue()
+
+                @Internal
+                Property<String> getUnused()
+            }
+
+            abstract class SomeTask extends DefaultTask {
+                @Nested
+                abstract Bean getBean()
+
+                @Nested
+                abstract Bean getUnusedBean()
+
+                @Internal
+                abstract Property<String> getValue()
+
+                @Internal
+                abstract Property<String> getUnused()
+
+                @TaskAction
+                void run() {
+                    println "this.value = " + value.getOrNull()
+                    println "this.unused = " + unused.getOrNull()
+                    println "this.bean.value = " + bean.value.getOrNull()
+                    println "this.bean.unused = " + bean.unused.getOrNull()
+                    println "this.unusedBean.value = " + unusedBean.value.getOrNull()
+                }
+            }
+
+            task ok(type: SomeTask) {
+                value = "42"
+                bean.value = "42"
+            }
+        """
+
+        when:
+        configurationCacheRun "ok"
+        configurationCacheRun "ok"
+
+        then:
+        outputContains("this.value = 42")
+        outputContains("this.unused = null")
+        outputContains("this.bean.value = 42")
+        outputContains("this.bean.unused = null")
+        outputContains("this.unusedBean.value = null")
+    }
+
+    def "task can reference itself"() {
+        buildFile << """
+            class SomeBean {
+                private SomeTask owner
+            }
+
+            class SomeTask extends DefaultTask {
+                private final SomeTask thisTask
+                private final bean = new SomeBean()
+
+                SomeTask() {
+                    thisTask = this
+                    bean.owner = this
+                }
+
+                @TaskAction
+                void run() {
+                    println "thisTask = " + (thisTask == this)
+                    println "bean.owner = " + (bean.owner == this)
+                }
+            }
+
+            task ok(type: SomeTask)
+        """
+
+        when:
+        configurationCacheRun "ok"
+        configurationCacheRun "ok"
+
+        then:
+        outputContains("thisTask = true")
+        outputContains("bean.owner = true")
+    }
+}
